Improved handling of factorio server settings
Settings inteded for the `server-settings.json` file are set under the `factorio_settings` variable in the role variables. This dict is just converted to a json object and stored in the server-settings.json file. This has eliminated the need for the server-settings.json.js2 template file, so it has been removed. Resolves #2.