More config file support
Added variables for settings values in the map-gen-settings.json, map-settings.json and server-whitelist.json files. All the variables are set to the defaults in the example files, but I've also added a variable to enable/disable the creation of these files (variable name + "_enaled", "factorio_server_whitelist_enabled" for example). This will need a little more work as the files aren't removed when the coresponding variable is set to false after previously being true.