Restart service on settings change
Added handlers to restart or reload the service if any of the config files change. The service is reloaded only if the systemd service file is changed. My own personal uses of the role have motivated me to add the `service_restart_permitted` variable which allows the service to be restarted when config files are applied. Setting this variable to `false` will prevent the service from being restarted if there are changes. This allows config changes to be made while the server is in use.

I have also slightly changed the way the factorio_server_settings dictionary is defined. This is to handle instances where a single key within the factorio_server_settings dictionary is defined by the user, which before would overwrite the default values, thus causing the server to fail to start. This now uses the jinja2 combine() method to merge the default values in `default_factorio_server_settings` with whatever is set by the user in `factorio_server_settings`. Documentation needs to be updated to very clearly state that values in the `default_factorio_server_settings` should not be partially overwritten.